Standardize on one convention for angles
The two conventions were that 0 was up and 90 was right, and that 0 was
right and 90 was up. The latter (when converted to radians) is used by
`psMat.skew`; Python's trigonometric functions use the former (converted
to radians). Everything now uses the former convention because
trigonometry is used more often than skewing.