Name glyphs by shape and anchors, not by character
It is possible for a single glyph to correspond to multiple characters,
or to multiple positional forms of the same character. Glyphs are now
named based on their shape (line, curve, or circle), angles, and
chirality, eliminating unnecessary duplicates. This will have more of an
effect with the introduction of semicircle and quarter-circle vowels.