Place circle vowels within curve consonants
A circle vowel should go within an adjacent curve consonant. If both
adjacent consonants are curves, it goes within the first.

Some glyphs have identical contour and anchor points, but have differing
behavior in downstream lookups. It is no longer sufficient to
distinguish such glyphs by the suffixes "nom" and "con", since two
identical final-form circle glyphs can be substituted with different
medial-form glyphs depending on whether the preceding consonant is a
curve. Therefore, glyphs are now suffixed with the name of the glyph
class responsible for the existence of the glyph. This is overkill, but
optimization can wait.