Ensure that circle vowels near curves get loops
If the presence of a curve consonant changes the chirality of a circle
vowel glyph, it is now represented as a `Circle` instead of a `Curve` to
ensure a loop; otherwise, it would look like one of the curve vowels.

Between some pairs of curves, e.g. U+1BC1A DUPLOYAN LETTER N and U+1BC1B
DUPLOYAN LETTER J, a circle vowel now appears to have two loops: one for
the vowel and one due to the overlap of the curves. In such cases, the
circle vowel should use `Curve`. This is hard to fix because if the
curves are of much different sizes, they will not overlap, so the vowel
should use `Circle` after all.

I am not actually sure about that, so I am postponing the issue.