Refactor lookups to use phases
Previously, all of the lookups were generated at once. This was
confusing and inefficient: confusing because the logic was all mixed
together in one function; inefficient because without any notion of
lookup ordering, there is no way to know which glyphs are valid inputs
to a lookup, so all glyphs must be accounted for.

The build script now uses phases. A phase represents a mapping between
sets of glyphs: in font terms, a sequence of lookups. Each phase is
passed the glyphs from the preceding phase, calculates the new set of
glyphs, emits one or more lookups to implement the mapping, and outputs
the new set of glyphs. (The first phase gets glyphs from the input code
points. This corresponds to the 'cmap' table.) This lets each lookup
focus on its possible input glyphs, ignoring those created only by
downstream lookups. The long-term plan is to use phases to build a
dependency DAG of lookups, allowing the detection and removal of
effectively duplicate glyphs, of which there are many.

For now, though, this is just a start. It does not even support
topographical features yet.