Add topographical phases and schema annotations
A schema represents a glyph plus extra build-time information.
Phases can add or remove annotations (arbitrary data) from schemas.
Currently, this is used for whether a code point is default ignorable
and the input and output contexts (if known). Future uses may include
whether a character can take a DTLS, the language system, and the
orientation convention for vowels. Basically, it is any information that
cannot be deduced from the glyph but which is relevant to lookups.

Unfortunately, this system is so general that the font has too many
duplicate glyphs and the 'curs' lookup breaks, so a couple schemas are
temporarily commented out.