Work around a FontForge .fea parsing bug
Given an OpenType feature file with a contextual substitution referring
to an anonymous multiple substitution, FontForge silently ignores all
but the first glyph of the output sequence, making it a single or
ligature substitution instead. It ought to make a multiple substitution
if possible, or, if there are multiple glyphs in the input sequence,
at least print a warning.