Implement a discretionary ligature
Discretionary ligatures are those which normally require a ZWJ, but do
not if 'dlig' is enabled. Only one such ligature is implemented: the
Pernin reversed ⟨ă⟩ for ⟨ăr⟩. Unfortunately, it was not possible to
implement the other Pernin reversed circles: that generates too many
glyphs and, once again, 'curs' stops working.

One of the 'ccmp' phases had to be changed to 'abvs': in the USE, 'ccmp'
is always applied before 'liga' or 'dlig', regardless of lookup order.
There is a good reason for this, but as Duployan benefits from neither
Indic reordering nor Arabic shaping, it is rather a nuisance.