Merge identical-looking and -behaving schemas
If two schemas represent glyphs with identical contour points, anchor
points, glyph classes, and widths (i.e. they are identical-looking), and
if there are no lookups that differentiate between them (i.e. they are
identical-behaving), then it is possible to compile them into one glyph.
This decreases the glyph count from 10806 to 9197 for the current font.
What a savings!

This replaces the hard-coded concept of schema identity.