Use Python 3.6
FontTools is scheduled to require Python 3.5 or above for new features
on 30 June 2019. The minimum for this project is now Python 3.6 because
it uses a couple of convenience features introduced in version 3.6.
Version 3.7 would have been preferable, because it guarantees that dict
iteration order is insertion order; however, because FontForge installs
its Python extensions in the global site-packages and Travis provides no
build environment whose default Python is 3.7, 3.6 will have to do.