Improve the rendering of circle vowels
If two non-orienting cochiral curves flank a circle vowel without
forming a straight angle, the circle vowel is now rendered as a curve
instead of a circle, because the adjacent curves cross and therefore
already form a loop. (This is not geometrically guaranteed, but it seems
to be a better approximation for real Duployan letters.)

I had thought that reversed circles should be rendered as circles inside
the angle, but they actually should be rendered outside the angle but
curving in the opposite direction.

There is still a rendering problem: if the reversed circle vowel's glyph
is a curve, the next glyph starts where the curve glyph ends, so the
glyphs on either side of the curve might not cross and the curve remain
open. A circle vowel must always produce some sort of loop.