Use FontForge 20190801
Instead of building and installing FontForge from source, Travis now simply
downloads the AppImage. This is much faster but requires many plumbing changes.
The AppImage runs Python 3.7, so Travis should too, which means `virtualenv
--system-site-packages` is no longer possible (because Travis doesn’t provide it
for Python 3.7 on Bionic) or necessary (because the `fontforge` and `psMat`
modules come with the AppImage). Running FontForge as an AppImage requires
setting the environment variable `PYTHONHOME` to get Python to run properly, and
using the environment variable `OWD` to get the current working directory
outside the image. The image was configured to run a GUI, which requires a
screen, which requires xvfb.

For development, I don’t use an AppImage, but the local build still needs to be
Python-in-FontForge instead of FontForge-in-Python, for consistency with CI.
That requires setting `PYTHONPATH` explicitly for some reason.