Only rerun a phase if feedback is possible
If a phase’s lookups’ rules don’t use backtracking sequences, no new
schema generated by the phase can ever have an effect on those rules.
It is only necessary to rerun a phase if feedback is possible, i.e. a
glyph produced in a chaining substitution could appear in the
backtracking sequence when processing the next glyph.