Use more Extension lookups & reenable all shading
Shaded curves and compound circle vowels took such a long time to
generate because FontTools was fixing offset overflows by promoting
lookups to Extension type one by one. There are hundreds of lookups, and
each round of promoting a lookup and trying again takes a couple
seconds. In the end, out of around 600 GSUB lookups, only about the
first 100 were not eventually promoted.

The fix is to, when a lookup overflows, pessimistically promote all
preceding lookups, not just one. It probably wastes some space in the
font file, but the speed-up is worth it.