Ban features that the phase system misanalyzes
The phase system assumes that every feature is either optional or is
applied to the whole buffer at once, without any special behavior.
Features for which that is not true are banned to avoid surprising bugs.

Features like 'rtlm', 'frac', 'rphf', and 'isol' are applied to specific
glyphs based on context and/or Unicode properties. The only ones of
interest for Duployan are related to fractions, which is a long way off.

'rand' applies one glyph at a time with a random feature value each
time. Since the builder never generates alternate substitutions, the
“random” value would always be 1, so it would be safe to allow 'rand'.
For future compatibility, though, in case that changes, it is best to
ban it.

'rphf' and 'pref' reorder glyphs.

Features like 'ccmp' apply one cluster at a time. Without reimplenting
USE clustering, it is unsafe to use them in ligature substitutions. It
is easier to avoid them entirely.

As it turns out, the only banned feature in use was 'ccmp' twice, both
in `dont_ignore_default_ignorables`. The second was a ligature
substitution, though in a context that was safe.