Restrict phases’ access to classes & named lookups
Previously, each phase got a fresh class dictionary, so it was unable to
access other phases’ classes, but it was still able to overwrite them
and to refer to them in rules. (The same is true for named lookups.)
Both problems are now fixed by adding a phase-specific prefix to every
key created, accessed, or referenced by a phase. There is a special
prefix for global keys.