Freeze classes that `add_rule` optimizes based on
`add_rule` includes two optimizations that depend on the glyphs in the
added rule. If an input schema was created in the current phase, the
glyph can’t exist yet, so the rule is skipped as unnecessary. Every
output schema is added to `output_schemas`. If any rule item is a class
instead of a glyph, the same optimization applies.

There is a potential problem. If a schema is later added to a class, it
might invalidate the decision to skip the rule, or the new schema might
never get added to `output_schemas`. Therefore, to be safe, such classes
are frozen: they can be used but not augmented.

All this also goes for named lookups, mutatis mutandis.