Tweak how schemas track their code points
`Schema.cp` is now `Schema.cmap` to emphasize that it should only be
used when updating 'cmap' for a new glyph. The value for an unmapped
glyph is `None`; the old value was `-1` just because that is what
FontForge did.

`Schema.cps` represents an empty string with an empty list, instead of
`[-1]`.

`Schema.sort_key` prefers glyphs mapped to code points in NFD. That only
affects U+2000 EN QUAD and U+2001 EM QUAD, which now lose out to U+2002
EN SPACE and U+2003 EM SPACE. The character name substitution from
“QUAD” to “SPACE” is therefore no longer necessary. (Of course, because
of a limitation in `sift_groups`, neither of these pairs is merged. But
if they ever do get merged, they will have the right names.)