Support sequences of orienting curves
A sequence of orienting curves is oriented as if only one of the curves
was present. If the sequence is initial, the chosen curve is the last in
the sequence; otherwise, the first curve is chosen. The other curves are
oriented solely in relation to the chosen curve, with alternating
chirality and with each curve rotated to avoid angles.

The only attestation of more than two in a row is a sequence of many
U+1BC46 DUPLOYAN LETTER I, where no length is precisely defined. This is
the origin of U+1BC50 DUPLOYAN LETTER YE, which, by contrast, may be a
zigzag and always follows the angle of F.