Apply secants only to roots, not child nodes
Unicode categorizes secants as letters, so a sequence of the form <root,
secant, U+1BCA0 SHORTHAND FORMAT LETTER OVERLAP, child> should represent
the child overlapping the secant. (This font does not support secants as
tree roots, but in principle the encoding allows for it.) However, in
that example, it is more likely that the secant was intended to apply to
the root; the only other possible order is <root, U+1BCA0, child,
secant>. Therefore, secants must ignore child nodes when determining
their bases.