Remove fixed-width breaking spaces
[UTN #37] says that fixed-width spaces preserve the Duployan baseline,
whereas the variable-width U+0020 and U+00A0 reset it to neutral.
However, only non-breaking fixed-width spaces are fit for this purpose.
If a line is broken at a fixed-width space, the following line will be
shaped in isolation from the first line; i.e. the baseline cannot be
preserved from one line to the next. The reading of a text would
therefore change depending on where its line breaks happened to fall.

[UTN #37]: https://www.unicode.org/notes/tn37/utn37-1-duployan.pdf

The only suitable fixed-width spaces are thus the non-breaking ones,
U+2007 FIGURE SPACE and U+202F NARROW NO-BREAK SPACE. U+2007 probably
doesn’t make sense in this context, but I left it in for future
consideration. If more than two baseline-preserving space widths are
significant, that requires either some new space characters in Unicode,
or a sequence of multiple U+202Fs.

The fixed-width spaces are sometimes used to represent clause or
sentence boundaries. In that role, I don’t think they are supposed to
preserve the baseline, so they might return to the font someday as
non-joining characters.