Suppress mark width adjustment on nonjoining bases
Otherwise, there would be a large space to the right of the base.
Technically, a wide diacritic on a narrow non-joining base could overlap
adjacent glyphs; this is the same as in Latin where ⟨i⟩ and ⟨î⟩ need
different kerning. The solution would be to calculate widths for all
glyphs at runtime, not just the joining ones. Perhaps it will happen one
day. In the meantime, it is unlikely to be a real problem.