Do not shade U+1BC4F DUPLOYAN LETTER LONG I
In some contexts, U+1BC4F is just a dot mark: there is no base to shade,
so U+1BC4F looks identical to <U+1BC4F, U+1BC9D>, which is confusing. I
could have changed `validate_shading` to account for that, but it didn’t
seem worth it. The only mode that uses both U+1BC4F and U+1BC9D is
Lockett’s Shorthand Alphabet (second edition). It does use a shaded
diagonal line for “i as in bide” but it is not clear that that should be
encoded as <U+1BC4F, U+1BC9D>, and anyway, Unicode doesn’t even support
that mode.