Invalidate overlaps before non-childable letters
Some letters can be parents but not children. If one such letter is in a
child position, the associated overlap control is invalid, but was
previously not being invalidated.

U+1BC30 DUPLOYAN LETTER J N and U+1BC31 DUPLOYAN LETTER J N S are now
such letters. (Previously, they could not be parents.) Changing these
two letters was mainly for the sake of some good regression tests, but
it can be rationalized because they are actually ligatures encoded
atomically. They should be treated as much as possible as sequences,
which means they can’t be children. (This is incidentally also why they
can’t be shaded.)