Improve certain Chinook Jargon abbreviations
The proper place for two letters two overlap depends on the shapes of
the letters. This font instead spreads the overlap points evenly along
the length of the letter and hopes that is good enough, but it isn’t.
Certain Chinook Jargon abbreviations are now supported slightly better
than before, but this is a targeted hack. This would ideally apply to
all letters.