Do not shade any circles
This makes the 'GSUB' table smaller, making room for more important
features. Circles really ought to be shadable, but most of them already
aren’t, and it might be confusing to only support shading for a subset.
The shaded glyphs for circles looked wrong anyway: they were shaded
along the entire stroke, whereas they should only be shaded on the
downstroke.