Add guidelines to the glyphs of separate affixes
Until separate affixes are properly supported, which might not happen
for the first release, it is better to give them guidelines, both to
show their intended positions relative to their bases and to emphasize
that this is not the correct rendering.

U+1BC81 DUPLOYAN AFFIX HIGH TIGHT ACUTE and U+1BC91 DUPLOYAN AFFIX LOW
TIGHT ACUTE are now placed at the same heights as their non-“tight”
counterparts. I can find no evidence that there is any such subtle
height distinction. The real distinction is in the horizontal position:
the Pernin prefix “cer-” is “an acute accent written from right to left,
slightly above and to the left of the sign following” whereas the Pernin
prefix “por-” is “above and to the right of the next sign”.