Make U+1BC9F perfectly horizontal
The misleadingly named U+1BC9F DUPLOYAN PUNCTUATION CHINOOK FULL STOP is
essentially a two-stroke dash. The code chart gives it a slanted glyph,
which is attested, but plenty of other attestations give it a horizontal
glyph. I am making it horizontal to emphasize its nature as a dash
instead of as some sui generis Chinook Jargon thing.