Orient medial circles regardless of any next curve
Between a line consonant and a curve consonant with no angle between
them, there are two ways to orient a circle letter. Previously, primary
orientation put the circle within the curve. Now, it puts the circle in
its final position, as if the curve were not there. This should make a
future IME more intuitive: users will be able to type a circle how it
looks directly, instead of predicting how an adjacent curve might flip
it. It also matches how the système Duployé codifié orients circles.