Let less prominent glyphs set the baseline
The system of hub markers is generalized to support an arbitrary number
of hub priority levels. The first hub at the most visually prominent
priority level in a stenogram defines the baseline. The levels (by
decreasing prominence) are:

1. Dotted guidelines and most non-orienting letters
2. Orienting letters and U+1BC01 DUPLOYAN LETTER X
3. U+1BC03 DUPLOYAN LETTER T and its cognates

Anything else (like secants and U+1BC00 DUPLOYAN LETTER H) or anything
following an overlap control is ignored for determining the baseline.