Make subantiparallel lines intersect around a loop
Subantiparallel lines are lines that are almost 180° apart. When an
orienting loop appears between them, they are, in practice, too short to
actually intersect. The solution is modify the angle of one of them.

This implementation is better than nothing but is still not perfect. The
full solution causes overflow in fontTools.