Add some simple compression for width numbers
A sequence of width digits in `Builder._add_width_markers` can be
represented by a single width number, which the next lookup expands into
its digits. This makes the first lookup a lot smaller at the expense of
a new rule in the second lookup. It is therefore optimal to compress
only widths that are frequent enough. Any width that appears more than
once is frequent enough.

Here is why. Every number has 7 digits. If a number is compressed, 7
glyph IDs are replaced with 1 glyph ID, saving 6 glyph IDs (12 bytes)
per occurrence of that width in the first lookup. In the second lookup’s
coverage table, the worst case is that it is format 2 and the new width
requires a new coverage range (6 bytes). The rule expanding the width to
its digits takes 7 glyph IDs (14 bytes). Overall, every compressed width
saves 12 × 𝑛 − 20 bytes. That is positive when 𝑛 > 5⁄3, and ⌈5⁄3⌉ = 2,
so compression is optimal for any width that appears 2 times or more.

To verify my calculations, I built the font with a variety of minimum
counts and measured the size of each font’s GSUB table.

| Count | Size ∕ B |
| ----: | -------: |
|     1 |  1675014 |
|     2 |  1671772 |
|     3 |  1674436 |
|     4 |  1678126 |
|     5 |  1682592 |
|     6 |  1688022 |
|     7 |  1694508 |
|    10 |  1709286 |
|    25 |  1767444 |
|    50 |  1898754 |
|   125 |  2093542 |
|   250 |  2236278 |
|   500 |  2342232 |
|  1000 |  2478460 |
|  1500 |  2518768 |
|  9999 |  2602598 |
| 99999 |  2764126 |

The minimum size corresponds to a count of 2.