Represent carry value 0 implicitly
Representing a carry value of 0 with a glyph is redundant: the font can
assume that if the carry value is not 1, it is 0. Generating all those
unnecessary glyphs made the `_sum_width_markers` lookup run afoul of
HarfBuzz’s `HB_BUFFER_MAX_LEN_FACTOR` and `HB_BUFFER_MAX_LEN_MIN` for
any input more than a few words long.