Remove the special behavior of U+1BC4F LONG I
The code chart says U+1BC4F DUPLOYAN LETTER LONG I is used for Pernin,
where it “angles like an ‘f’ when adjacent to a k-type consonant”, but
this is not exactly how it works in Pernin, and other modes do it
differently. It is best to reinterpret this character as a simple,
non-orienting line.

In Pernin’s Universal Phonography, the long I is usually a small K. It
can also be a small tick in other directions, which may be doubled back
on itself. It may have a dot. If there is a dot and the adjacent letters
form an angle, there might not be anything of the long I besides the
dot. It isn’t as standardized as the code chart implies, and different
manuals show different behavior.

In Perrault-Duployan, it is a small K or L, without a dot.

In Wagner Phonography, it is a small K or L, or a dot outside an angle.

Various modes use small versions of letters. To support them, it will be
necessary to encode many small letters. We might as well use U+1BC4F for
the small K. The aforementioned positional or discretionary variants of
the long I can use other small letters. The dot can use an affix
character; it’s not really an affix, but it acts like a combining
character that goes between two bases and represents its own sound, so
it is like what the current encoding calls an affix.

It is better to do it this way because a graphetic model is easier to
work with. It ensures that all fonts will render the same strings
compatibly and minimizes the confusion from the code chart’s current
recommendation. In general, when a character takes multiple contextual
glyphs that do not orient continuously but instead switch between a few
fixed forms, it is better to encode each form as a different code point,
because the switching rules are always slightly different between modes.