Don’t alternate chirality in orienting sequences
Adjacent cochiral orienting curves are attested in the word “shisyuit”
(‘Jesuit’) in line 4 of [page 26 of _Kamloops Wawa_, September 1904]. It
is therefore inappropriate to force adjacent orienting curves to be
antichiral. That means that the sequence of many orienting semicircles
representing “y” in various modes must be represented by U+1BC46 or
U+1BC47 (to define the overall direction) followed by many U+1BC47.

[page 26 of _Kamloops Wawa_, September 1904]: https://open.library.ubc.ca/viewer/kwawa/1.0082063#p31z-3r0f:

The sequence <U+1BC46, U+1BC46> doesn’t make sense and is probably
unattested. If it is attested, it probably looks like a larger
semicircle containing a smaller semicircle, but for now the font uses
the wrong but most straightforward rendering: a circle.