Do not flip curves to avoid overlap
A basic tenet of Duployan is that orienting curves can flip to avoid
overlap; for example, the vowel in “tim” must be flipped to make “tem”.
However, a small amount of overlap is allowed, and some letters’ entry
and exit angles differ depending on the mode, so it can be ambiguous
whether a certain string contains overlap. Some cases are indeed
unambiguous, but it is not clear where to draw the line.

Therefore, this commit introduces a new encoding model. Primary
orienting curves turn counterclockwise in general and clockwise in
initial position. Secondary orienting curves turn the other way. There
are no exceptions. A word that is pronounced “tim” but looks like “tem”
must be encoded as “tem”.

This is another step towards a more graphetic model.