Remove U+202F from Noto Sans Duployan
Noto fonts are supposed to work in Chrome. Chrome has a bug where U+202F
NARROW NO-BREAK SPACE resets the baseline. I infer that it does not
support cursive joining for spaces. In any case, that makes this
character pointless in the Noto font. I could keep it for the sake of
non-buggy applications like Firefox, but it is better to keep the
behavior consistent.

The non-Noto build still includes U+202F.