Change the encoding model of orienting sequences
A circle as the leader of an orienting sequence keeps its inherent
chirality. Previously, it was eligible to change, using the same
“outside the angle” logic used for monophthongs. For example, in
<U+1BC04 F, U+1BC41 A, U+1BC46 I>, the A was counterclockwise, but
appending U+1BC03 T made the A become clockwise. This was an unintuitive
complication that did not match what is most natural for Chinuk Pipa.

The ideal implementation would remove the concept of dependent schemas
that are ignored for topography, now that leading orienting glyphs no
longer need to ignore anything. Unfortunately, I don’t have time to do
such a major refactoring so close to the deadline. Maybe later.