Shift curves’ exit points instead of adding U+25CC
The angles of curve consonants are not the same across all modes, so a
string that would cause overlap in one mode might not in another. It is
therefore inappropriate to insert dotted circles to interrupt sequences
that are perfectly fine in some modes. In particular, the Chinook Jargon
word “𛰛𛱇𛰙” was getting a dotted circle.

The solution is to shift the exit points of these potentially
overlapping curves to make them less likely to overlap for realistic
inputs. They may still overlap for unrealistic inputs, which is a shame,
but at least they still look bogus: the problem that the dotted circles
were meant to solve was that a curve might overlap in such a way as to
look like a circle letter, which it no longer would.