Build in a virtual environment
If the build step shares the same environment as the other steps, the
`matrix.requirements == 'oldest'` job might end up downgrading a
dependency if dev-requirements.txt and requirements.txt have different
minimum versions for the same dependency. This is currently the case
with fontTools, where the oldest version for dev is 4.28.4 but the
oldest version for building normally is 4.28.2; luckily, this particular
discrepancy does not break anything.