Test subsetted fonts
Sometimes a smaller version of the font is necessary, as in [Go Noto
Universal]. The subsetted font removes the width system, overlap trees,
and topographical forms. The letters are not cursively joined but it is
still fairly intelligible. The feature 'rclt' is excluded from the
subset, so some lookups have had their features reassigned.

[Go Noto Universal]: https://github.com/satbyy/go-noto-universal