Make the width system work with script tag DFLT
The default shaper zeroes mark glyphs’ widths after GPOS, unlike USE
which zeroes them before GPOS. That makes it pointless to modify the x
advances of mark glyphs like the final left, right, and anchor bound
markers. The fix is to make them not be mark glyphs.

That fix breaks the cursive attachment of the final letter to `_.END`.
The solution to that is to remove the anchor point from `_.END` and add
it to all of `_.RDX.*E0`.