Create carry in 1st `sum_width_markers` iteration
A carry marker is necessary in the width system regardless of the radix
(assuming it is an integer greater than or equal to 2, which is a safe
assumption) because it is always possible to make a list of digits whose
sum is greater than or equal to the radix, which causes carrying. It is
therefore safe to create a carry marker in the first iteration of
`sum_width_markers` without checking that carrying actually happens.

Previously, none the rules in the first iteration were contextual, since
there was no carry marker available yet. To make sure the lookup was
nevertheless considered contextual, the backtracking context was an
instance of `_AlwaysTrueList` to make sure every rule appeared to have
a non-empty context. The first iteration now produces contextual rules,
so this hack has been removed.