Allow shading for Sloan secants U+1BC76 & U+1BC77
Sloan-Duployan uses a shaded secant for “exer”. The code point for its
secant is either U+1BC76 or U+1BC77, or it might depend on context. For
maximum compatibility, both can now be shaded. The other secants still
cannot be shaded, since that is not attested and it makes the font
unnecessarily bigger.