Minimize the rotation of initial WA/WO/WOW
Chinook Jargon generally prefers the inner circle to stay at or near the
top of the outer circle. The outer circle rotates only just enough to
let the inner circle form a satisfying loop, which I have arbitrarily
defined as a curve of at least 180°.

Some of this is probably true in medial and final position too, but this
is a complex topic, so I am only addressing initial position for now.