Fix an error in `cast_cff_number` for large floats
If the input to `cast_cff_number` was a float with an integral value
outside CFF’s integer range (a signed 32-bit integer), it would end up
passing the `int` version of it to `encodeIntCFF`, which would raise a
`struct.error`. The range check now accounts for float inputs and avoids
the exception.