Fail if no old version satisfies all specifiers
get-old-requirements.py has a bug. If a certain version of a dependency
has a transitive dependency with conflicting specifiers, it discards the
latter dependency. It should instead backtrack and try a different
version of the former dependency. Until it is fixed properly, this
commit makes get-old-requirements.py fail with an error message, to make
it obvious that manual intervention is required.