Support superscript & subscript digit code points
The glyphs were already supported via 'sups' and 'subs' but now the
glyphs are in 'cmap' too. This is primarily to support fractions in
Phonography Made Easy, where superscript 1 and 2 represent thirds and
subscript 1 and 3 represent quarters. The other superscript and
subscript digits do not appear in that mode but are supported here for
consistency.

Superscript and subscript digits are attested in at least Romanian in
contexts where they would be used in longhand, i.e. dimensional
exponents and chemical formulas. Markup is probably most appropriate for
those contexts, but the dedicated code points are very convenient.