Add U+E009 COMBINING RING-AND-DOT ABOVE for “‰”
The combination of the ring and dot is attested in _Stenografie după
sistemul Duployé_, p. 173. It should be encoded with a single code point
instead of as <U+030A, U+0307> because it means something different than
the combination of those two characters’ meanings (“%” and “× 100”
respectively) and because [L2/23-206R] says “The SAH has noted in the
past that it is inadvisable to require special stacking rules for
diacritics, so pairs of adjacent diacritics \[...] should be encoded
individually.” Specially stacking rules might be appropriate for
Duployan marks, but this is a common-script mark used only on digits.

[L2/23-206R]: https://www.unicode.org/L2/L2023/23206r-harrington-diacritics.pdf