Avoid accidental butt caps
When an open contour doubles back on itself, overlapping perfectly,
FontForge strokes that part of the contour with a butt cap instead of
the requested cap. This is probably a bug in FontForge, but it is easy
to work around it here by setting `skip_drawing` for certain components
that would otherwise overlap.