Make medial reversed `Wa` more handwritinglike
It is not perfect: in some sequences, the inner circle looks too big, or
the outer circle is not closed by the inner circle. In handwriting,
these flaws are addressed by squishing the outer circle. That might be
necessary in this font, but it might look odd, because the rest of the
font is so uncompromisingly self-consistent and geometric.