Separate subantiparallels less indiscriminately
It is only necessary to separate subantiparallel lines when there is
something visually between them, like an orienting vowel before or after
the pair of lines that curves inwards, or when a loop between them might
not be closed. This commit makes lines be separated only when needed by
deferring `separate_subantiparallel_lines` till after the topographical
phases. As an unfortunate side effect, ticks are now inserted between
cognate lines, even when the separation makes them not cognate, the
solution to which might be to split tick insertion from topography into
its own phase.