Fix sequences of three or more circles
They were broken by b42983a3059dac6522336d0004dc09db6ded9ce8. That
commit’s optimization of using a reverse chaining substitution is still
a good idea, so the solution is not to revert it; rather, it is to
remove a different optimization. Like many phases, the buggy one,
`join_circle_with_adjacent_nonorienting_glyph`, used the following
optimization to effectively run only once:

```python
    if len(original_schemas) != len(schemas):
        return [lookup]
```

That is kind of a hack because the phase system runs each phase until
the set of schemas reaches a fixed point. Returning early on non-initial
iterations does obviously produce a fixed point, but it goes against the
spirit of the thing. It should only be used to save computation time if
running an extra iteration would have no effect. However, nothing
actually verifies that it would have no effect, and in this case, that
led to a bug.

The real solution is therefore to make the phase correctly process
schemas that were created in previous iterations. Returning early
happened to work when the phase produced a normal chaining substitution,
but it was always a risky optimization.