In CI, eliminate `skip-check` and add `build-venv`
The original point of `skip-check` was to build with only the basic
requirements, not the dev requirements, in case the latter had an effect
on the output font. Testing is a dev activity so it had to be skipped.

Now, everything in CI is done in a virtual environment. There is one for
building and one for testing. In some jobs, those are actually the same
environment; this verifies that building in a dev environment works. In
other jobs, they are distinct environments; this verifies that none of
the dev requirements are actually basic requirements. The build
artifacts come from the latter jobs. The matrix variable `build-venv`
controls whether there is one virtual environment or two. Either way,
every job runs tests.