Add U+0023 NUMBER SIGN, U+20B6 LIVRE TOURNOIS SIGN
They are both used as symbols for francs in contexts like “1#50” in [_Le
Sténographe_]. As with other currency symbols, they are not really
Duployan characters, but they are common-script characters attested in
purely Duployan contexts, which is enough to include them in this font.
Furthermore, they were easy and fun to draw.

[_Le Sténographe_]: https://books.google.fr/books?id=tZhcvveMiq8C

I’m not certain about U+20B6 LIVRE TOURNOIS SIGN. There are many
attestations of a one-stroke franc sign, so it is clearly an intentional
glyph, not a mistake for a two-stroke U+0023 NUMBER SIGN. U+20B6 would
be semantically appropriate, except that the livre tournois had been
long obsolete by 1870. Maybe people were still using its symbol in
handwriting. Just to be safe, I’ve excluded it from the Noto build.