Remove the Arabic-Indic digits
Benaroya’s Ottoman Turkish stenography is based on both Duployan and
Prévost-Delaunay and should not be unified with Duployan. This font
should therefore not include any characters included only because of
Benaroya’s system or the modern TBMM system derived from it.

Gontard’s Ottoman Turkish system is Duployan, but the manual does not
give any examples of digits in shorthand.

U+1BC00 DUPLOYAN LETTER H may still be shaded. Although that feature was
added because of attestations in Turkish, lots of modes use shading, so
it is plausible that a Duployan mode uses a shaded dot.