Orient medial hook relative to following letter
Medial hooks contrast with normal semicircles in Ellis’s mode.

In Sloan-Duployan, a semicircle between a non-curve and a curve is
oriented relative to the curve. There may be an angle against the
preceding non-curve, but it is not required. Orienting semicircles in
Sloan-Duployan should therefore be encoded as U+1BC7A or U+1BC7B when
visually appropriate, but they don’t represent different letters from
U+1BC47 and U+1BC46.