Add U+0312 COMBINING TURNED COMMA ABOVE
[A Duployan transcription] of the Lord’s Prayer in Chinese in Sichuan
(“在天我等父者” etc.) uses turned apostrophes to mark aspiration, as was
also common in romanizations of Chinese before pinyin. It is written
above and to the left of the consonant letter. The proper encoding is
not clear. Is it a combining mark, a Duployan affix, or a punctuation
mark? If it is a combining mark, does it go above the base or above and
to the right? Of all available assigned characters, U+0312 seems to be
the most appropriate.

[A Duployan transcription]: http://caneco.my.coocan.jp/001g/2/22.html