Rename the font to “Rawnd Musmus Duployan”
The testing font is now “Ilo Snas Duployan”. “Rawnd musmus ilo snas” is
Chinook Jargon for ‘round cow without rain’, as in a spherical cow in a
vacuum, alluding to this project’s goal of being a technical proof of
concept, and not necessarily a beautiful and perfectly accurate
emulation of handwriting.

The font name suffix “Unjoined” is now “Uncow”, alluding to the Chinook
Jargon word for ‘untied’ and to the “Musmus” earlier in the name.